Updates in UserController , update-role.html. Delete, getAllUsers
